VisitTimeShift And Companion Added

+++ ManagementAPI/Mappers/CustomerProfileMapper.cs
@@ -12,7 +12,7 @@ public class CustomerProfileMapper : Profile
     {
         CreateMap<CreateCustomerRequestDto, Customer>()
             .ForMember(dest => dest.Id, opt => opt.Ignore())
-            .ForMember(dest => dest.Status, opt => opt.MapFrom(x => (short)Status.Active))
+            .ForMember(dest => dest.Status, opt => opt.MapFrom(x => Status.Active))
             .ForMember(dest => dest.CreatedOn, opt => opt.MapFrom(x => DateTime.Now))
             .ForMember(dest => dest.CreatedById, opt => opt.MapFrom(x => 1));
